Guns updated
Added several Common gun Archetypes

Pistol - A basic Pistol with a standard fire rate and middle-of-the-road damage.

AR - Assault rifle medium damage and full auto with a decent Magazine size.

SMG - Submachine gun low damage but makes up for it with a very high rate of fire and a large magazine size.

Rifle - like a hunting rifle. high damage and fast bullet speed with an increased life time so that the bullet travels further.

Sniper - Highest damage but slowest fire rate. make up for the low rate of fire with bullet penetration. Can take out multiple enemies if they are lined up

I'm still working on a burst fire gun I might be able to figure it out before turn-in time but we will see.

I Made several changes to the GunManager and Gun2 scripts to increase the ease of gun creation.

Creating new guns is now as easy as cloning an existing gun prefab and changing the exposed variables in the Inspector.